// src/app/interfaces/task.ts
export interface Task {
  _id:         ID;
  files:       any[];
  subtasks:    Subtask[];
  categories:  string[];
  title:       string;
  description: string;
  deadline:    string;
  priority:    number;
  status:      number;
  owner:       string;
  updated_at:  string;
  created_at:  string;
  comments:    Comment[];
  members:     Member[];
}

export interface CreateTask{
  files:       any[];
  subtasks:    Subtask[];
  categories:  string[];
  title:       string;
  description: string;
  deadline:    string;
  status:      number;
  user_ids:    string[];
}


export interface ID {
  $oid: string;
}

export interface Member {
  email: string,
	uid: string,
	first_name: string,
	last_name: string,
	allow_password_change: boolean,
	provider: string,
	locker_locked_at: boolean,
	locking_name: boolean,
  id: string,
}// Generated by https://quicktype.io

export interface Comment {
  _id:        ID;
  content:    string;
  created_at: string;
  user:       UserComment;
}

export interface UserComment {
  first_name: string;
  last_name:  string;
  id:         ID;
}

export interface Subtask {
  title:  string;
  status: number;
  subtask_id: string
  
}